Fix implicit conversion warnings

+++ src/Binance/Signer.cpp
@@ -113,8 +113,8 @@ std::vector<uint8_t> Signer::aminoWrap(const std::string& raw,
         if (prefixWithSize) {
             cos.WriteVarint64(contentsSize);
         }
-        cos.WriteRaw(typePrefix.data(), typePrefix.size());
-        cos.WriteRaw(raw.data(), raw.size());
+        cos.WriteRaw(typePrefix.data(), static_cast<int>(typePrefix.size()));
+        cos.WriteRaw(raw.data(), static_cast<int>(raw.size()));
     }
 
     return std::vector<uint8_t>(msg.begin(), msg.end());
